# Speed up `plug_leaks`