@nnethercote 2016-10-03
BTW, I used Valgrind's "DHAT" profilter to find this. It told me that lots of malloc calls were happening in and beneath `plug_leaks`.