@arielb1 2016-10-03
We should also be doing this for `leak_check`.