@nnethercote 2016-10-03
> We should also be doing this for leak_check.

`leak_check` doesn't show up nearly as high as `plug_leaks` in profiles, but it does show up. I'll do it in a follow-up. Thank you for the suggestion.