@arielb1 2016-10-04
leak-check done by #36931.