@nnethercote 2016-10-04
> leak-check done by #36931.

Huh. I have a local patch to add an early return to `leak_check`. It produces a measurable speed-up of a couple of percent on some of the rustc-benchmarks, but it doesn't look anything like the commits in #36931.