@arielb1 2016-10-04
> Huh. I have a local patch to add an early return to `leak_check`. It produces a measurable speed-up of a couple of percent on some of the rustc-benchmarks, but it doesn't look anything like the commits in #36931.

How do you run the rustc benchmarks? I've been measuring bootstrap times. My patch short-cuts the fast path to `leak_check` at the callsites.