@nnethercote 2016-10-04
@arielb1: I updated my local repo to get #36931 and the speed-ups from my local patch have now evaporated, so I guess its effect overlapped with the effect of your commit. My patch was simpler, though, adding just a `if skol_map.is_empty() { return Ok(()); }` near the start of `leak_check`. I don't know the code well enough to tell if that's functionally equivalent to your commit.

As for running the benchmarks, I compared two stage1 compilers that were both configured with `--enable-optimize --enable-debuginfo`. I used the `compare.py` script for this, see https://internals.rust-lang.org/t/recent-rustc-benchmarks-improvements/4153 for instructions.