@arielb1 2016-10-04
It should be functionally equivalent. Is it performance-equivalent?