@arielb1 2016-10-04
The `leak_check` patch *seems* to be slower. Will investigate more tomorrow.