@nnethercote 2016-10-08
Good news -- the improvement for jld-day15-parser showed up clearly on perf.rust-lang.org.
![plug_leaks](https://cloud.githubusercontent.com/assets/1940286/19215948/fd415de6-8df7-11e6-8291-aa9ca7ce037e.png)
It's the big drop on the RHS on Oct 4. The time dropped from 1.402 to 1.199, which is a 1.169x speed-up (a.k.a a 14.5% reduction in execution time).

The other benchmarks mentioned in the first comment also saw speedups similar to what I measured locally, which is good.